# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "operationId": "products_list",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Busca en nombre, SKU, código de barras y marca",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "stock",
            "required": false,
            "in": "query",
            "description": "Semáforo de existencias",
            "schema": {
              "type": "string",
              "enum": [
                "available",
                "low",
                "out"
              ]
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Solo activos (true) o solo dados de baja (false)",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isRetail",
            "required": false,
            "in": "query",
            "description": "Solo los que se venden a clientas",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 200,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: products.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista el catálogo con su semáforo de existencias",
        "tags": [
          "Productos"
        ]
      },
      "post": {
        "operationId": "products_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: products.create"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un producto",
        "tags": [
          "Productos"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "products_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: products.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar un producto",
        "tags": [
          "Productos"
        ]
      },
      "patch": {
        "operationId": "products_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: products.update"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un producto",
        "tags": [
          "Productos"
        ]
      },
      "delete": {
        "operationId": "products_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: products.delete"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Da de baja un producto. Exige que no queden existencias",
        "tags": [
          "Productos"
        ]
      }
    },
    "/api/v1/products/{id}/restore": {
      "post": {
        "operationId": "products_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: products.restore"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restaurar un producto dado de baja",
        "tags": [
          "Productos"
        ]
      }
    },
    "/api/v1/inventory/kardex": {
      "get": {
        "operationId": "inventory_kardex",
        "parameters": [
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "batchId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PURCHASE_IN",
                "SALE_OUT",
                "SERVICE_CONSUMPTION",
                "ADJUSTMENT",
                "RETURN_IN",
                "WASTE_OUT",
                "INITIAL_STOCK"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 200,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KardexPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: inventory.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Libro de movimientos. Es la fuente de verdad del inventario",
        "tags": [
          "Inventario"
        ]
      }
    },
    "/api/v1/inventory/batches": {
      "get": {
        "operationId": "inventory_batches",
        "parameters": [
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "onlyAvailable",
            "required": false,
            "in": "query",
            "description": "Solo lotes con existencias",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 200,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: inventory.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista los lotes por producto",
        "tags": [
          "Inventario"
        ]
      }
    },
    "/api/v1/inventory/alerts": {
      "get": {
        "operationId": "inventory_alerts",
        "parameters": [
          {
            "name": "withinDays",
            "required": false,
            "in": "query",
            "description": "Ventana de caducidad, en días",
            "schema": {
              "minimum": 1,
              "maximum": 365,
              "default": 30,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockAlertsEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: inventory.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Qué hay que pedir y qué está a punto de caducar",
        "tags": [
          "Inventario"
        ]
      }
    },
    "/api/v1/inventory/receive": {
      "post": {
        "operationId": "inventory_receive",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiveStockEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: inventory.receive"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registra una entrada de mercancía, con lote si el producto lo lleva",
        "tags": [
          "Inventario"
        ]
      }
    },
    "/api/v1/inventory/consume": {
      "post": {
        "operationId": "inventory_consume",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumeStockEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: inventory.adjust"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Descuenta existencias repartiendo por FEFO entre los lotes disponibles",
        "tags": [
          "Inventario"
        ]
      }
    },
    "/api/v1/inventory/adjust": {
      "post": {
        "operationId": "inventory_adjust",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustStockEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: inventory.adjust"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajusta existencias por recuento o merma",
        "tags": [
          "Inventario"
        ]
      }
    },
...
  },
  "info": {
    "title": "AppSalonBelleza API",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token obtenido en POST /api/v1/auth/login"
      }
    },
    "schemas": {
...
      "ProductPageResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaResponse"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "ProductEnvelopeResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProductResponse"
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "maxLength": 50,
            "example": "SH-ARG-500"
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "example": "Champú de argán 500 ml"
          },
          "brand": {
            "type": "string",
            "maxLength": 80
          },
          "barcode": {
            "type": "string",
            "maxLength": 80
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "example": 120.5
          },
          "costPrice": {
            "type": "number",
            "minimum": 0,
            "example": 65
          },
          "taxRate": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "example": 12,
            "description": "IVA aplicable. Por defecto, el 12 % de Guatemala"
          },
          "reorderPoint": {
            "type": "number",
            "minimum": 0
          },
          "reorderQuantity": {
            "type": "number",
            "minimum": 0
          },
          "unit": {
            "enum": [
              "UNIT",
              "MILLILITER",
              "LITER",
              "GRAM",
              "KILOGRAM"
            ],
            "type": "string"
          },
          "isRetail": {
            "type": "boolean",
            "default": true
          },
          "isInternal": {
            "type": "boolean",
            "default": false
          },
          "trackStock": {
            "type": "boolean",
            "default": true
          },
          "tracksBatches": {
            "type": "boolean",
            "default": false,
            "description": "Trazabilidad por lote y caducidad. Obligatoria en cosmética con caducidad"
          }
        },
        "required": [
          "sku",
          "name",
          "price"
        ]
      },
...
      "KardexPageResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KardexEntryResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaResponse"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}